FINALLY managed to revisit my code and add comments.
The purpose of this code is to take the greatest prime factor p of each
integer smaller than N. Then, add up every p. The gods revealed to me a
way to do this, but since I was coding by the inspiration of the muses,
I didn't have time to make any comments. But now that I have revisited
my code, after several months, I once again understand what the code
does. This is the algorithm that was revealed to me by the gods:

Make a list of every integer, from 0 to N. We're going to loop through
every prime number up to the square root of N. Let's start with 2.
Besides every multiple of 2, write the number 2. Then we move on to the
next prime, which is 3. Once again, besides every multiple of 3, we
write the number 3. But this time around, we might need to overwrite
several of the 2's we wrote earlier. Then move on to the next prime, and
the next, until we have used every prime up to the square root of N.
Overwriting is good, because it means that we found a greater prime
factor than the previous one.

At this point, we have taken care of the multiples of the primes which
are smaller than the square root of N. But there still remain the
multiples of the primes which are greater than the square root; we left
them blank. So every blank entry must be a prime number (provided it is
greater than 1). Hence, we take each blank entry j, and write the number
j next to it. We also overwrite every multiple of j, which are not that
many.

And that's it! That's the gist of it. What's left should be a list of
every integer from 0 to N, and besides each integer is its greatest
prime factor.

Well, almost. We still need to take care of the special cases of zero
and one. For the purposes of Project Euler 642, 0 is the greatest prime
factor of both zero and one, since we don't need to include them in our
final sum.

Then we take our list of greatest prime factors, and add them all up!

FIN.